// types/tmdb/credits.ts
export type Credits = {
	id: number;
	cast: Cast[];
	crew: Crew[];
};

export type Cast = {
	adult: boolean;
	gender: number;
	id: number;
	known_for_department: string;
	name: string;
	original_name: string;
	popularity: number;
	profile_path: string;
	credit_id: string;
	cast_id: number;
	character: string;
	order: number;
};

export type Crew = {
	adult: boolean;
	gender: number;
	id: number;
	known_for_department: string;
	name: string;
	original_name: string;
	popularity: number;
	profile_path: string;
	credit_id: string;
	department: string;
	job: string;
};